+++ prisma/migrations/20211014141636_add_reset_otp/migration.sql
@@ -1,0 +1,6 @@
+-- AlterTable
+ALTER TABLE "User" ADD COLUMN "resetPasswordOtp" INTEGER;
+
+-- RedefineIndex
+DROP INDEX "User.email_unique";
+CREATE UNIQUE INDEX "User_email_key" ON "User"("email");