+++ src/schema/schema.graphql.ts
@@ -1,55 +1,59 @@
 import { gql } from 'apollo-server'
 
 export default gql`
-  type User {
+  input CreateDraftInput {
     email: String!
+    content: String
+    title: String!
+  }
+
+  input ResetPasswordInput {
+    password: String!
+    passwordConfirmation: String!
+    code: Int!
+  }
+
+  input ForgotPasswordInput {
+    email: String
+  }
+
+  type AuthPayload {
+    jwt: String
+    user: User
+  }
+
+  type Post {
     id: ID!
+    title: String!
+    content: String
+    published: Boolean!
+    author: User
+  }
+
+  type User {
+    id: ID!
+    email: String!
     name: String
     posts: [Post!]!
   }
 
-  type Post {
-    author: User
-    content: String
-    id: ID!
-    published: Boolean!
-    title: String!
-  }
-
   type Query {
+    login(email: String!): String
     me: User
     feed: [Post!]!
-    filterPosts(searchString: String): [Post!]!
-    post(where: PostWhereUniqueInput!): Post
+    filterPosts(searchTerm: String): [Post!]!
+    post(id: Int!): Post
   }
 
   type Mutation {
-    createDraft(authorEmail: String, content: String, title: String!): Post!
-    deleteOnePost(where: PostWhereUniqueInput!): Post
-    publish(id: ID): Post
-    signupUser(data: UserCreateInput!): User!
-  }
+    #auth
+    resetPassword(input: ResetPasswordInput!): AuthPayload
+    forgotPassword(input: ForgotPasswordInput!): Boolean
 
-  input PostWhereUniqueInput {
-    id: ID
-  }
-
-  input UserCreateInput {
-    email: String!
-    id: ID
-    name: String
-    posts: PostCreateManyWithoutPostsInput
-  }
-
-  input PostCreateManyWithoutPostsInput {
-    connect: [PostWhereUniqueInput!]
-    create: [PostCreateWithoutAuthorInput!]
-  }
-
-  input PostCreateWithoutAuthorInput {
-    content: String
-    id: ID
-    published: Boolean
-    title: String!
+    createDraft(input: CreateDraftInput!): Post!
+    deleteOnePost(id: Int!): Post
+    publish(id: Int!): Post
+    # @TODO: add createUser mutation
+    #signup(input: UserCreateInput!): User!
   }
 `