+++ src/utils/index.ts
@@ -1,2 +1,5 @@
+export * from './email'
+export * from './helper'
 export * from './jwt'
 export * from './prisma'
+export * from './template'