+++ src/utils/jwt.ts
@@ -1,15 +1,21 @@
 import { compare, hash } from 'bcryptjs'
 import { sign, verify } from 'jsonwebtoken'
 import { clone } from 'lodash'
+import { Context } from './prisma'
 
 interface Token {
   id: string
 }
 
+type User = {
+  id: number
+}
+
 // get user id from auth token
-export function getUserId({ headers }: any) {
-  if (headers) {
-    const Authorization = headers.authorization || headers.Authorization
+export function getUserId({ request }: Context): string | undefined {
+  if (request && request.headers) {
+    const Authorization =
+      request.headers.authorization || request.headers.Authorization
 
     if (Authorization) {
       const token = Authorization.replace('Bearer ', '')
@@ -26,27 +32,8 @@ export function getUserId({ headers }: any) {
 }
 
 // issue new token based on payload
-export const issue = (payload: any, jwtOptions = {}) => {
-  return sign(
-    clone(payload.toJSON ? payload.toJSON() : payload),
-    process.env.JWT_SECRET as string,
-    jwtOptions,
-  )
-}
-
-export const hashPassword = (password: string): Promise<string | null> => {
-  return new Promise((resolve, reject) => {
-    if (!password || isHashed(password)) {
-      resolve(null)
-    } else {
-      hash(`${password}`, 10, (err, hash) => {
-        if (err) {
-          return reject(err)
-        }
-        resolve(hash)
-      })
-    }
-  })
+export const issue = (payload: string | User | Buffer, jwtOptions = {}) => {
+  return sign(clone(payload), process.env.JWT_SECRET as string, jwtOptions)
 }
 
 export const isHashed = (password: string) => {
@@ -55,6 +42,21 @@ export const isHashed = (password: string) => {
   }
 
   return password.split('$').length === 4
+}
+
+export const hashPassword = (password: string): Promise<string | null> => {
+  return new Promise((resolve, reject) => {
+    if (!password || isHashed(password)) {
+      resolve(null)
+    } else {
+      hash(`${password}`, process.env.SALT as string, (err, hash) => {
+        if (err) {
+          return reject(err)
+        }
+        resolve(hash)
+      })
+    }
+  })
 }
 
 export const validatePassword = (password: string, hash: string) => {