+++ CHANGELOG.md
@@ -1,27 +1,68 @@
-# [1.2.0](https://github.com/2wce/prisma-template/compare/v1.1.0...v1.2.0) (2021-10-08)
-
+# [1.7.0](https://github.com/2wce/prisma-template/compare/v1.6.0...v1.7.0) (2021-10-14)
 
 ### Features
 
-* upgrade packages ([2b5ba2a](https://github.com/2wce/prisma-template/commit/2b5ba2a69edf1ec7978c70b6d8045b675348c2a5))
+- add reset password input helper ([cc05ab2](https://github.com/2wce/prisma-template/commit/cc05ab22c99b7424992d0b159ba92d5eb1dcdfe3))
+
+# [1.6.0](https://github.com/2wce/prisma-template/compare/v1.5.0...v1.6.0) (2021-10-14)
+
+### Bug Fixes
+
+- fix deep source issues ([e19f490](https://github.com/2wce/prisma-template/commit/e19f4909e49b73e20a72d5424901e9e692733a9a))
+- fix login import ([eeadace](https://github.com/2wce/prisma-template/commit/eeadacec5fa864668f3736c68a34c9194b51ba8c))
+
+### Features
+
+- add reset flow mutations ([505b97f](https://github.com/2wce/prisma-template/commit/505b97f1c774b574e7d633456c483fbcd5daafc8))
+- add user fields ([362da7c](https://github.com/2wce/prisma-template/commit/362da7cdc2605cee6d1cb79275ddaec81e258ff6))
+
+# [1.5.0](https://github.com/2wce/prisma-template/compare/v1.4.0...v1.5.0) (2021-10-14)
+
+### Features
+
+- add login query ([6287112](https://github.com/2wce/prisma-template/commit/62871129f4f2bb7d314f0aaf196f75db25fe6484))
+
+# [1.4.0](https://github.com/2wce/prisma-template/compare/v1.3.0...v1.4.0) (2021-10-14)
+
+### Features
+
+- cleanup schema & remove ambiquous inputs ([fcbf285](https://github.com/2wce/prisma-template/commit/fcbf2855b835acb34e4fac056b696dd98ef95b88))
+- remove signup for now ([5b8f366](https://github.com/2wce/prisma-template/commit/5b8f366f7a1e3bb7c77548cde04e584390200f19))
+
+# [1.3.0](https://github.com/2wce/prisma-template/compare/v1.2.1...v1.3.0) (2021-10-14)
+
+### Features
+
+- remove any type ([1e81c51](https://github.com/2wce/prisma-template/commit/1e81c5126ebfa03b1945c0d01d30777c633569dc))
+- setup graphql codegen ([41acf6a](https://github.com/2wce/prisma-template/commit/41acf6a41667468e6e656771a02530415d5056f8))
+
+## [1.2.1](https://github.com/2wce/prisma-template/compare/v1.2.0...v1.2.1) (2021-10-14)
+
+### Bug Fixes
+
+- declare function before use ([ab13b8e](https://github.com/2wce/prisma-template/commit/ab13b8e6a4415e3665b36c52eb88ac999a0e97b8))
+
+# [1.2.0](https://github.com/2wce/prisma-template/compare/v1.1.0...v1.2.0) (2021-10-08)
+
+### Features
+
+- upgrade packages ([2b5ba2a](https://github.com/2wce/prisma-template/commit/2b5ba2a69edf1ec7978c70b6d8045b675348c2a5))
 
 # [1.1.0](https://github.com/2wce/prisma-template/compare/v1.0.0...v1.1.0) (2021-08-10)
 
-
 ### Features
 
-* change to gql schema ([eed4147](https://github.com/2wce/prisma-template/commit/eed414742255f870ac779839a4bbb07078e2ff85))
-* update tests ([64bdba8](https://github.com/2wce/prisma-template/commit/64bdba8b5db007bc8a34884a61900251c1c0f68c))
-* upgrade prisma & apollo ([f8a5a5a](https://github.com/2wce/prisma-template/commit/f8a5a5a0493f6a165a929403f4e6f2fa1f064c69))
+- change to gql schema ([eed4147](https://github.com/2wce/prisma-template/commit/eed414742255f870ac779839a4bbb07078e2ff85))
+- update tests ([64bdba8](https://github.com/2wce/prisma-template/commit/64bdba8b5db007bc8a34884a61900251c1c0f68c))
+- upgrade prisma & apollo ([f8a5a5a](https://github.com/2wce/prisma-template/commit/f8a5a5a0493f6a165a929403f4e6f2fa1f064c69))
 
 # 1.0.0 (2021-03-11)
 
-
 ### Features
 
-* add db helper scripts ([0896bea](https://github.com/2wce/prisma-template/commit/0896beae5fcc7d23eb5a897089b39de1d465359d))
-* add migration ([5a9731d](https://github.com/2wce/prisma-template/commit/5a9731d87fce0702850225b32ca92b3703800488))
-* add semantic release ([f44ad05](https://github.com/2wce/prisma-template/commit/f44ad05c4c7c74c12db010948457359a196a0069))
-* setup apollo server ([7a83e64](https://github.com/2wce/prisma-template/commit/7a83e64112d08b2bbe59fede1ae100374e83a694))
-* update dependencies ([e7cde24](https://github.com/2wce/prisma-template/commit/e7cde24059f8dac7058c8ddd189022e4db20aba3))
-* update deps ([d159206](https://github.com/2wce/prisma-template/commit/d15920699235ea155ca548d7594a9bea8c3da509))
+- add db helper scripts ([0896bea](https://github.com/2wce/prisma-template/commit/0896beae5fcc7d23eb5a897089b39de1d465359d))
+- add migration ([5a9731d](https://github.com/2wce/prisma-template/commit/5a9731d87fce0702850225b32ca92b3703800488))
+- add semantic release ([f44ad05](https://github.com/2wce/prisma-template/commit/f44ad05c4c7c74c12db010948457359a196a0069))
+- setup apollo server ([7a83e64](https://github.com/2wce/prisma-template/commit/7a83e64112d08b2bbe59fede1ae100374e83a694))
+- update dependencies ([e7cde24](https://github.com/2wce/prisma-template/commit/e7cde24059f8dac7058c8ddd189022e4db20aba3))
+- update deps ([d159206](https://github.com/2wce/prisma-template/commit/d15920699235ea155ca548d7594a9bea8c3da509))